refactor: Split file module into separate loader components with FileLoader trait
- Created generic FileLoader trait for extensibility (CSV, TOML, YAML, etc.)
- Split lazy_loader.rs (422 lines) into focused modules:
  - loaders/ndjson.rs (134 lines) - NDJSON file loader
  - loaders/json_array.rs (226 lines) - JSON array loader
  - loaders/single.rs (87 lines) - Single value loader
  - loaders/mod.rs (179 lines) - FileLoader trait and common types
- All loaders implement FileLoader trait with Item = Value
- LazyJsonFile enum implements FileLoader trait
- Maintained backward compatibility via lazy_loader.rs re-exports
- Updated all imports to use new loaders module
- Added documentation example for future loader implementations

Benefits:
- Single responsibility principle for each loader
- Format-agnostic trait allows future CSV/TOML/YAML loaders
- Improved testability and maintainability
- No functionality changes or regressions

Closes #18
