+++ src/file/mod.rs
@@ -1,2 +1,3 @@
 pub mod detect_file_type;
 pub mod lazy_loader;
+pub mod loaders;
